# api/openapi.yaml
openapi: 3.1.0
info:
  title: PlatformFoundry API
  description: |
    PlatformFoundry Internal Developer Platform API.

    This API provides programmatic access to all platform operations including:
    - Platform and workload management
    - Deployment orchestration
    - Resource provisioning
    - Observability and metrics
    - Policy and compliance
  version: 1.0.0
  contact:
    name: PlatformFoundry Team
    url: https://platformfoundry.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.platformfoundry.io/v1
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /platforms:
    get:
      summary: List platforms
      operationId: listPlatforms
      tags: [Platforms]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: List of platforms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformList'
    post:
      summary: Create platform
      operationId: createPlatform
      tags: [Platforms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformCreate'
      responses:
        '201':
          description: Platform created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '400':
          $ref: '#/components/responses/BadRequest'
  /platforms/{name}:
    get:
      summary: Get platform
      operationId: getPlatform
      tags: [Platforms]
      parameters:
        - $ref: '#/components/parameters/platformName'
      responses:
        '200':
          description: Platform details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update platform
      operationId: updatePlatform
      tags: [Platforms]
      parameters:
        - $ref: '#/components/parameters/platformName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformUpdate'
      responses:
        '200':
          description: Platform updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
    delete:
      summary: Delete platform
      operationId: deletePlatform
      tags: [Platforms]
      parameters:
        - $ref: '#/components/parameters/platformName'
      responses:
        '204':
          description: Platform deleted
...
  /events:
    get:
      summary: List events
      operationId: listEvents
      tags: [System]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
    platformName:
      name: name
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
      required: [error]
...
    Platform:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
        spec:
          type: object
        status:
          type: object
    PlatformCreate:
      type: object
      properties:
        name:
          type: string
        spec:
          type: object
      required: [name]
    PlatformUpdate:
      type: object
      properties:
        spec:
          type: object
    PlatformList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Platform'
        total:
          type: integer
...
    Event:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        source:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
    EventList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    Metadata:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...